+++ packages/frontend/src/components/common/Table/ScrollableTable/TableBody.tsx
@@ -11,7 +11,8 @@ import React, { FC } from 'react';
 import { getColorFromRange, readableColor } from '../../../../utils/colorUtils';
 import { getConditionalRuleLabel } from '../../Filters/FilterInputs';
 import { ROW_HEIGHT_PX, Tr } from '../Table.styles';
-import { TableContext, useTableContext } from '../TableProvider';
+import { TableContext } from '../TableProvider';
+import { useTableContext } from '../TableProvider/useTableContext';
 import BodyCell from './BodyCell';
 
 const VirtualizedArea: FC<{ cellCount: number; padding: number }> = ({
