+++ packages/frontend/src/components/common/Table/ScrollableTable/TableFooter.tsx
@@ -2,7 +2,7 @@ import { isNumericItem } from '@lightdash/common';
 import { flexRender } from '@tanstack/react-table';
 import React from 'react';
 import { FooterCell } from '../Table.styles';
-import { useTableContext } from '../TableProvider';
+import { useTableContext } from '../TableProvider/useTableContext';
 
 const TableFooter = () => {
     const { table, data, footer } = useTableContext();
