+++ packages/frontend/src/components/common/Table/ScrollableTable/index.tsx
@@ -1,6 +1,6 @@
 import { FC, useRef } from 'react';
 import { Table, TableScrollableWrapper } from '../Table.styles';
-import { useTableContext } from '../TableProvider';
+import { useTableContext } from '../TableProvider/useTableContext';
 import TableBody from './TableBody';
 import TableFooter from './TableFooter';
 import TableHeader from './TableHeader';
