+++ packages/frontend/src/components/common/Table/TableProvider/useTableContext.tsx
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { Context, TableContext } from '../TableProvider';
+
+export function useTableContext(): TableContext {
+    const context = useContext(Context);
+    if (context === undefined) {
+        throw new Error('useTableContext must be used within a TableProvider');
+    }
+    return context;
+}
